Return full info to region.
(i.e. return a row for every fish, not just those with NMFS_DNA_IDS.  I also update the bycatch_IDs to correct the duplicate issues with 131/137/120/149.

We had to modify the DISSECTION_BARCODE_NMFS in the reconciled meta data file to make this all work out.  That has now been done.  All fish are now accounted for.